requirement.txt
aim==3.17.5
aiohttp==3.8.4
flake8==6.1.0
fqdn==1.5.1
g2pK==0.9.4
hydra-core==1.3.2
isoduration==20.11.0
jsonpointer==2.4
librosa==0.10.0.post2
pip==23.1.2
pipdeptree==2.13.0
python-dotenv==1.0.0
pytorch-lightning==2.0.5
torchaudio==0.13.1
uri-template==1.3.0
webcolors==1.13
